feat(debug) Do not check incoming signature if both env variable and config option are set
