+++ lib/events/proxy.js
@@ -42,7 +42,7 @@ const _isAxiosError = err => err && err.response && err.response.status
  * @example express.use('/events', etvas.events())
  */
 const createProxyHandler = () => async (req, res) => {
-  if (!req.body) {
+  if (!req.body || typeof req.body !== 'object') {
     res
       .status(500)
       .json({ error: 'Invalid implementation: missing json body parser' })
@@ -51,12 +51,7 @@ const createProxyHandler = () => async (req, res) => {
     )
   }
 
-  const signature = req.get('x-etvas-signature')
-  if (!signature) {
-    return res
-      .status(400)
-      .json({ error: 'Invalid request: missing signature header' })
-  }
+  const { suppressSignatureCheck } = config.get('debug') || {}
 
   const { name, payload, timestamp } = req.body
   // Always have a name
@@ -64,24 +59,32 @@ const createProxyHandler = () => async (req, res) => {
     return res.status(500).json({ error: 'Empty name' })
   }
 
-  const now = Date.now()
-  const oneMinute = 60000
-  const timeLapse = config.get('eventTimeLapse') || oneMinute
-  if (
-    isNaN(timestamp) ||
-    typeof timestamp !== 'number' ||
-    timestamp <= 0 ||
-    Math.abs(now - timestamp) > timeLapse
-  ) {
-    return res.status(400).json({ error: 'Event timestamp invalid' })
-  }
+  if (suppressSignatureCheck !== true || !process.env.ETVAS_SDK_DEBUG) {
+    const signature = req.get('x-etvas-signature')
+    if (!signature) {
+      return res
+        .status(400)
+        .json({ error: 'Invalid request: missing signature header' })
+    }
 
-  const hmacKey = config.get('eventSecret')
+    const now = Date.now()
+    const oneMinute = 60000
+    const timeLapse = config.get('eventTimeLapse') || oneMinute
+    if (
+      isNaN(timestamp) ||
+      typeof timestamp !== 'number' ||
+      timestamp <= 0 ||
+      Math.abs(now - timestamp) > timeLapse
+    ) {
+      return res.status(400).json({ error: 'Event timestamp invalid' })
+    }
 
-  const canonical =
-    typeof req.body === 'string' ? req.body : JSON.stringify(req.body)
-  if (!hmac.verify(canonical, signature, hmacKey)) {
-    return res.status(403).json({ error: 'Invalid signature' })
+    const hmacKey = config.get('eventSecret')
+
+    const canonical = JSON.stringify(req.body)
+    if (!hmac.verify(canonical, signature, hmacKey)) {
+      return res.status(403).json({ error: 'Invalid signature' })
+    }
   }
 
   // do we have custom handler ?
