+++ lib/init/init.js
@@ -10,7 +10,8 @@ module.exports = options => {
       apiKey: options.apiKey,
       eventSecret: options.eventSecret,
       productVariants: options.productVariants,
-      reqSignatureSecret: options.reqSignatureSecret
+      reqSignatureSecret: options.reqSignatureSecret,
+      debug: options.debug
     },
     'init'
   )
